requirements.txt
Flask>=2.0
gunicorn
openai==0.28.0
python-dotenv
 